services/account_data_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-pubsub==2.18.4
pymongo==4.6.0
python-dotenv==1.0.0
pydantic==2.5.0
